requirements.txt
fastapi
pydantic
google-cloud-aiplatform
google-auth
uvicorn
